Move import/export code to separate client file
